examples/deepagents/requirements.txt
deepagents>=0.0.7
langchain-openai>=0.2.0
pyyaml>=6.0.0
